# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Parking Lot LObby",
    "description": "PLL Api Documentation",
    "termsOfService": "",
    "contact": {
      "name": "matrix marketer",
      "url": "",
      "email": ""
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.1"
  },
...
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
          "type": "apiKey",
          "name": "Authorization",
          "in": "header"
      }
    },
    "schemas": {
      "Login":{
        "type":"object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
          
        },
        "required": [
          "email",
          "password"
        ]
      },
      "Register":{
        "type":"object",
        "properties": {
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "password": {
              "type": "string"
          },
          "repeatPassword": {
          "type": "string"
          },
          "agreeTermCondition": {
          "type": "boolean",
          "default":false
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password",
          "repeatPassword",
          "agreeTermCondition"
        ]
      },
      "Forgot":{
        "type":"object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPassword": {
        "type":"object",
        "properties": {
          "password": {
            "type": "string"
          },
          "repeatPassword": {
            "type": "string"
          },
          "resetToken": {
            "type": "string",
            "description":"This token was sent to email which was provied into forgot Api"
          }
        },
        "required": [
          "password",
          "repeatPassword",
          "resetToken"
        ]
      },
      "BusinessInformation":{
        "type":"object",
        "properties": {
          "companyName": {
            "type": "string"
          },
          "companyAddress": {
            "type": "string"
          },
          "companyNumber": {
            "type": "string"
          },
          "timezone": {
              "type": "string"
          },
          "language": {
          "type": "string"
          }                  
        },
        "required": [
          "companyName",
          "companyAddress",
          "companyNumber"                  
        ]
      },
      "AlertSettings":{
        "type": "object",
        "properties": {
          "confirmationAlert": {
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":false
              },
              "message":{
                "type": "string"
              }
            }
          },
          "nextInLineAlert": {
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":false
              },
              "message":{
                "type": "string"
              }
            }
          },
          "checkInAlert": {
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":false
              },
              "message":{
                "type": "string"
              }
            }
          },
          "checkOutAlert": {
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":false
              },
              "message":{
                "type": "string"
              }
            }
          },
          "noShowAlert":{
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":false
              },
              "message":{
                "type": "string"
              }
            }
          },
          "parkingSpotAlert":{
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":false
              },
              "message":{
                "type": "string"
              }
            }
          }
        },
        "required": [
          "confirmationAlert",
          "nextInLineAlert",
          "checkInAlert",
          "checkOutAlert",
          "noShowAlert",
          "parkingSpotAlert"
        ]
      },
      "AdditionalSettings":{
        "type": "object",
        "properties": {
          "desktopAdditional": {
            "type": "object",
            "properties": {
              "is_active":{
                "type": "boolean",
                "default":true
              },
              "is_timer":{
                "type": "boolean",
                "default":true
              },
              "is_exit":{
                "type": "boolean",
                "default":true
              },
              "is_checkIn":{
                "type": "boolean",
                "default":true
              },
              "is_delayed":{
                "type": "boolean",
                "default":true
              }
            }
          },
          "inforClientPositionLine":{
            "type": "boolean",
            "default":true
          }
        },
        "required": [
          "desktopAdditional",
          "inforClientPositionLine"          
        ]
      },
      "StylingSetting":{
        "type":"object",
        "properties":{
          "stylingScreen":{
            "type": "boolean",
            "default":false
          }
        },
        "required": [
          "stylingScreen"
        ]
      },
      "ClientInformationSettings":{
        "type": "object",
        "properties": {
          "clientInformation": {
            "type": "object",
            "properties": {
              "name":{
                "type": "boolean",
                "default":true
              },
              "is_required":{
                "type": "boolean",
                "default":true
              },
              "firstLastName":{
                "type": "boolean",
                "default":true
              }
            }
          }
        },
        "required": [
          "clientInformation"
        ]
      },
...
      "MarkMessageRead":{
        "type":"object",
        "properties": {
          "patientId": {
            "type": "string",
            "required":true
          }
        },
        "required": [
          "patientId"
        ]
      },
      "UpdatePatientProfile":{
        "type":"object",
        "required": ["patientId","first_name","last_name","fullNumber"],
        "properties": {
          "patientId": {
            "type": "string"
          },
          "first_name": {
            "type": "string"
          },
          "last_name": {
            "type": "string"
          },
          "fullNumber": {
            "type": "string"
          }
        }        
      },
...
    },
...
  },
  "paths": {
    "/user/login": {
        "post": {
          "summary": "Login",
          "description": "This for login a user",
          "tags": [
            "Basic Profile"
          ],
          "responses": {
            "$ref": "#/components/responses"
          },
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Login"
                }
              }
            }
          }
        }
    },
    "/user/register": {
        "post": {
          "summary": "Register",
          "description": "This for register a user",
          "tags": [
            "Basic Profile"
          ],
          "responses": {
            "$ref": "#/components/responses"
          },
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Register"
                }
              }
            }
          }
        }
    },
    "/user/forgot": {
        "post": {
          "summary": "forgot",
          "description": "This for send a email to generate new password",
          "tags": [
            "Basic Profile"
          ],
          "responses": {
            "$ref": "#/components/responses"
          },
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref":"#/components/schemas/Forgot"
                }
              }
            }
          }
        }
    },
    "/user/reset-password": {
      "post": {
        "summary": "For reset password using token from email",
        "description": "Here we can reset password by using token which was send into email from forgot Api",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/ResetPassword"
              }
            }
          }
        }
      }
  },
    "/user/settings": {
      "get": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "fetch clinic(provider) settings",
        "description": "This for fetch clinic settings.",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        }
      }
    },
    "/user/updateBusinessInformation": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "update Business Information",
        "description": "This for Update a clinic(provider) basic Information settings",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/BusinessInformation"
              }
            }
          }
        }
      }
    },
    "/user/updateAlertSettings": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "update Alert Settings",
        "description": "This for Update a clinic(provider) Alert settings",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertSettings"
              }
            }
          }
        }
      }
    },
    "/user/updateAdditionalSettings": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "update Additional Settings",
        "description": "This for Update a clinic(provider) Additional Information settings",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdditionalSettings"
              }
            }
          }
        }
      }
    },
    "/user/updateStylingSettings": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "update Styling Settings",
        "description": "This for Update a clinic(provider) Styling settings",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StylingSetting"
              }
            }
          }
        }
      }
    },
    "/user/updateclientInformationSettings": {
      "post": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "update Additional Settings",
        "description": "This for Update a clinic(provider) Additional Information settings",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClientInformationSettings"
              }
            }
          }
        }
      }
    },
    "/user/getNotifications": {
      "get": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "fetch Notifications for clinic",
        "description": "This for fetch clinic notifications.",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        }
      }
    },
    "/user/markMessageRead": {
      "put": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "mark all message for patient as read",
        "description": "This for mark read for messages.",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/MarkMessageRead"
              }
            }
          }
        }
      }
    },
...
    "/user/updatePatientInfo" : {
      "put": {
        "security":[
          {
            "ApiKeyAuth": []
          }
        ],
        "summary": "Update profile",
        "description": "This for update patient name or number.",
        "tags": [
          "Basic Profile"
        ],
        "responses": {
          "$ref": "#/components/responses"
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref":"#/components/schemas/UpdatePatientProfile"
              }
            }
          }
        }
      }
    },
...
  }
}